# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WTD Lead Generation API",
    "description": "Workspace routes require X-API-Key. Microsoft login is not required for research or editing. Approving, sending and resolving email require a connected Microsoft mailbox via X-Session-Token.",
    "version": "3.1.0"
  },
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "Public Settings",
        "operationId": "public_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BackendKey": []
          }
        ]
      }
    },
...
    "/api/v1/leads/{lead_id}": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "Get Lead",
        "operationId": "get_lead_api_v1_leads__lead_id__get",
        "security": [
          {
            "BackendKey": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Contact": {
        "properties": {
          "person_id": {
            "type": "string",
            "title": "Person Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name",
            "default": ""
          },
          "last_name": {
            "type": "string",
            "title": "Last Name",
            "default": ""
          },
          "job_title": {
            "type": "string",
            "title": "Job Title",
            "default": ""
          },
          "email": {
            "type": "string",
            "title": "Email",
            "default": ""
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "default": ""
          },
          "management_level": {
            "type": "string",
            "title": "Management Level",
            "default": ""
          },
          "accuracy_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy Score"
          }
        },
        "type": "object",
        "required": [
          "person_id"
        ],
        "title": "Contact"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntentSignal": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": ""
          },
          "signal_score": {
            "type": "integer",
            "title": "Signal Score",
            "default": 0
          },
          "audience_strength": {
            "type": "string",
            "title": "Audience Strength",
            "default": ""
          },
          "signal_date": {
            "type": "string",
            "title": "Signal Date",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "topic"
        ],
        "title": "IntentSignal"
      },
...
      "LeadView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "website": {
            "type": "string",
            "title": "Website",
            "default": ""
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "default": 0
          },
          "tier": {
            "type": "string",
            "title": "Tier",
            "default": ""
          },
          "score_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Score Reasons"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/IntentSignal"
            },
            "type": "array",
            "title": "Signals"
          },
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/Contact"
            },
            "type": "array",
            "title": "Contacts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "run_id",
          "company_id",
          "company_name"
        ],
        "title": "LeadView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BackendKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "MicrosoftMailbox": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Token"
      }
    }
  }
}